pyproject.toml
[project]
name = "arcgis-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arcgis>=2.4.2",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
